fpo-api/requirements.txt
# Main
django>=3.0.3  # 2.2 increases minimum sqlite version, which is not supported on the centos image
django-filter>=2.2,<2.3
djangorestframework>=3.11.0,<4
djangorestframework-bulk>=0.2.1,<1
django-cors-headers>=3.4.0,<4
git+https://github.com/impak-finance/django-oidc-rp.git@e99534729c9a2fa9cd8e62cd50d9916409376038#egg=django-oidc-rp

# Database managment
psycopg2-binary>=2.7.3.1,<3

# Web Server
gunicorn>=19.7.1,<20
whitenoise>=4.1.2,<4.2

# Documentation
django_rest_swagger>=2.1.2,<3

# Testing
pytest-cov>=2.5.1,<3
# pytest-django>=3.1.2,<4

# Util
requests>=2.18.4,<3
pypdf2>=1.26.0,<2
jsonfield>=3.1.0,<4

# for loading custom fonts
# XXX: This causes incompatibility errors
# WeasyPrint>=51
